+++ packages/react/test-app/tests/e2e/specs/overlay-components/IonModal.cy.ts
@@ -34,3 +34,45 @@ describe('IonModal', () => {
     cy.get('ion-modal div').contains('overriden value')
   });
 });
+
+describe('IonModal: conditional rendering', () => {
+
+  // Issue: https://github.com/ionic-team/ionic-framework/issues/25590
+  it('should push a new IonCard when dismissed', () => {
+    cy.visit('/overlay-components/modal-conditional-sibling');
+    // Renders a card before and after the modal
+    cy.get('ion-card').should('have.length', 2);
+
+    cy.get('ion-button').click();
+
+    cy.get('ion-card').should('have.length', 4);
+  });
+
+  // Issue: https://github.com/ionic-team/ionic-framework/issues/25590
+  it('should be conditionally rendered', () => {
+    cy.visit('/overlay-components/modal-conditional');
+
+    cy.get('ion-modal').should('not.exist');
+    cy.get('ion-button#renderModalBtn').click();
+
+    cy.get('ion-modal').should('be.visible');
+
+    cy.get('ion-button#dismissModalBtn').click();
+    cy.get('ion-button#renderModalBtn').should('be.visible');
+  });
+
+  it('should display an inline modal within a modal', () => {
+    cy.visit('/overlay-components/modal-datetime-button');
+
+    cy.get('ion-modal').should('not.exist');
+    cy.get('ion-button#renderModalBtn').click();
+
+    cy.get('ion-modal').should('be.visible');
+
+    cy.get('ion-datetime-button').click();
+
+    cy.get('ion-modal#datetimeModal').should('be.visible');
+    cy.get('ion-datetime').should('be.visible');
+  });
+
+});