+++ packages/vue/src/components/IonIcon.ts
@@ -1,7 +1,6 @@
 import { h, defineComponent } from 'vue';
 import { isPlatform } from '@ionic/core/components';
-import { defineCustomElement } from '../utils';
-import { IonIcon as IonIconCmp } from 'ionicons/components/ion-icon.js';
+import { defineCustomElement } from 'ionicons/components/ion-icon.js';
 
 export const IonIcon = /*@__PURE__*/ defineComponent({
   name: 'IonIcon',
@@ -18,7 +17,7 @@ export const IonIcon = /*@__PURE__*/ defineComponent({
     src: String
   },
   setup(props, { slots }) {
-    defineCustomElement('ion-icon', IonIconCmp);
+    defineCustomElement();
     return () => {
       const { icon, ios, md } = props;